broke up app.js into workable parts. sessions not working.
